refactor: modularize tx history parsing, merge, and cursor helpers
